ajustements pour linux 10

+++ noethys/Dlg/DLG_Nbre_inscrits.py
@@ -114,7 +114,10 @@ class Renderer_gauge(object):
         
         if etat == "alerte" :
             tailleImage = 16
-            dc.DrawBitmap(wx.Bitmap(Chemins.GetStaticPath("Images/16x16/Attention.png"), wx.BITMAP_TYPE_ANY) , largeurGauge-tailleImage-2, int((h-tailleImage)/2))
+            dc.DrawBitmap(wx.Bitmap(Chemins.GetStaticPath("Images/16x16/Attention.png"),
+                                    wx.BITMAP_TYPE_ANY) ,
+                          int(largeurGauge-tailleImage-2),
+                          int((h-tailleImage)/2))
 
     def SetValeurs(self, nbrePlacesPrises=0, nbrePlacesDispo=0, nbrePlacesAttente=0):
         self.nbrePlacesPrises = nbrePlacesPrises
